requirements.txt
AdvancedHTMLParser==8.0.1
beautifulsoup4==4.6.3
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
cssutils==1.0.2
Django==1.11.13
django-appconf==1.0.2
django-googledrive-storage==1.3.4
django-inlinecss==0.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
EasyProcess==0.2.3
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.11.3
idna==2.7
mock==2.0.0
oauth2client==4.1.3
pbr==4.2.0
ply==3.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.6.4
pynliner3==0.6
python-dateutil==2.7.3
pytz==2018.5
PyVirtualDisplay==0.2.1
QueryableList==3.1.0
requests==2.19.1
rsa==4.0
selenium==3.14.0
six==1.11.0
slimit==0.8.1
uritemplate==3.0.0
urllib3==1.23
django-currentuser==0.2.3

